chore: Print the string associated with the message method
